change script to ts. Remove unused imports. Use i18n instead of react-i18n
