+++ packages/visual-editor/package.json
@@ -44,8 +44,8 @@
     "docs:preview": "vitepress preview .",
     "generate-registry": "tsx src/components/registry/build-registry.cts",
     "build-registry-docs": "pnpm run docs:build && pnpm run generate-registry",
-    "i18n:insert": "pnpm exec ts-node scripts/insertI18n.js",
-    "i18n:insert:dry": "pnpm exec ts-node scripts/insertI18n.js --dry-run",
+    "i18n:insert": "pnpm exec tsx scripts/insertI18n.ts",
+    "i18n:insert:dry": "pnpm exec tsx scripts/insertI18n.ts --dry-run",
     "i18n:scan": "i18next-scanner --config i18next-scanner.config.cjs"
   },
   "dependencies": {
@@ -73,12 +73,10 @@
     "lz-string": "1.5.0",
     "next-themes": "^0.3.0",
     "picocolors": "^1.0.1",
-    "prettier": "^3.3.3",
     "prompts": "^2.4.2",
     "pure-react-carousel": "^1.32.0",
     "react-color": "^2.19.3",
     "react-error-boundary": "^5.0.0",
-    "react-i18next": "^15.5.2",
     "react-phone-number-input": "^3.4.12",
     "sonner": "^1.5.0",
     "tailwind-merge": "^2.4.0",
@@ -127,9 +125,8 @@
     "libphonenumber-js": "^1.11.19",
     "lint-staged": "^15.2.7",
     "minimist": "^1.2.8",
-    "node-fetch": "^3.3.2",
-    "p-limit": "^6.2.0",
     "playwright": "^1.51.1",
+    "prettier": "^3.3.3",
     "react-icons": "^5.2.1",
     "semver": "^7.6.3",
     "tailwindcss": "^3.4.6",
