+++ packages/visual-editor/scripts/test.ts
@@ -1,0 +1,1 @@
+console.log("Hello from TS");
